feat: new template for routing and server
